requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.3
urllib3==2.3.0
